+++ docs/.eleventy.js
@@ -1,0 +1,16 @@
+module.exports = (eleventyConfig) => {
+  eleventyConfig.addPassthroughCopy({
+    'src/_includes/styles': 'styles',
+  })
+  eleventyConfig.addPassthroughCopy({
+    'src/_includes/static': 'static',
+  })
+
+  return {
+    passthroughFileCopy: true,
+    htmlTemplateEngine: 'njk',
+    dir: {
+      input: 'src'
+    },
+  }
+}